refactor: move job endpoints under /declared/task router
